Added function for search and moved cross and len function to utils.
